+++ app/models/game.go
@@ -14,8 +14,3 @@ type Game struct {
 	CreatedAt   time.Time `json:"created_at"`
 	UpdatedAt   time.Time `json:"updated_at"`
 }
-
-// TableName 指定表名
-func (Game) TableName() string {
-	return "games"
-}